+++ Controller/WebPayment/ReturnFromPaymentGateway.php
@@ -31,10 +31,8 @@ class ReturnFromPaymentGateway extends Action
         try {
             $this->paylinePaymentManagement->synchronizePaymentWithPaymentGatewayFacade($this->getToken(), true);
         } catch (\Exception $e) {
+            $this->loggerPayline->critical(__METHOD__, ['token'=>$this->getToken(), 'exception'=>['message'=>$e->getMessage(), 'code'=>$e->getCode()]]);
             $this->messageManager->addErrorMessage($e->getMessage());
-            $this->loggerPayline->critical(__CLASS__. ' : ' .__FUNCTION__);
-            $this->loggerPayline->critical('Token # '.$this->getToken());
-            $this->loggerPayline->critical($e->getMessage());
             $isSuccess = false;
         }