+++ Controller/WebPayment/ReturnFromWidget.php
@@ -31,9 +31,7 @@ class ReturnFromWidget extends Action
         try {
             $this->paylinePaymentManagement->synchronizePaymentWithPaymentGatewayFacade($this->getToken(), true);
         } catch (\Exception $e) {
-            $this->loggerPayline->critical(__CLASS__. ' : ' .__FUNCTION__);
-            $this->loggerPayline->critical('Token # '.$this->getToken());
-            $this->loggerPayline->critical($e->getMessage());
+            $this->loggerPayline->critical(__METHOD__, ['token'=>$this->getToken(), 'exception'=>['message'=>$e->getMessage(), 'code'=>$e->getCode()]]);
             $this->messageManager->addErrorMessage($e->getMessage());
             $isSuccess = false;
         }